refactor: change name to some method

+++ app/src/services/admin.py
@@ -1,13 +1,18 @@
+from flask import abort
 from src.models import UserRecord, Admin 
 from src.firebase import admin_sdk
 from src.models import Privilegies
 
 
+
 class AdminService:
 
-    def get_user(self, uid):
+    def get(self, uid):
         user_record = UserRecord.get_user(uid, app=admin_sdk)
         user = Admin.query.filter_by(uid=user_record.uid).first()
+
+        if not user_record or not user:
+            abort(404, description='NotFound', response='not_found')
 
         return {
             'uid': user_record.uid,
@@ -15,13 +20,12 @@ class AdminService:
             'b': user
         }
 
-    def list_user(self, page=1):
-        per_pages = 10
-        users = Admin.query.paginate(page, per_pages, error_out=False)
-        
+    def list(self, page=1, per_pages=10):
+        users = Admin.query.paginate(page, per_pages or 10, error_out=False)
+       
         return users
 
-    def create_user(self, body):
+    def create(self, body):
         try:
             user_record = UserRecord(
                     email=body['email'],
@@ -52,9 +56,12 @@ class AdminService:
         except KeyError as ex:
             abort(400, description='BadRequest', response=str(ex))
 
-    def update_user(self, uid, body):
+    def update(self, uid, body):
         user_record = UserRecord.get_user(uid, app=admin_sdk) 
         user = Admin.query.filter_by(uid=user_record.uid).first()
+
+        if not user_record or not user:
+            abort(404, description='NotFound', response='not_found')
 
         user_record.serialize(body)
         user_record.update_user()
@@ -73,9 +80,12 @@ class AdminService:
             'b': user
         }
 
-    def update_field_user(self, uid, body):
+    def update_field(self, uid, body):
         user_record = UserRecord.get_user(uid, app=admin_sdk) 
         user = Admin.query.filter_by(uid=user_record.uid).first()
+
+        if not user_record or not user:
+            abort(404, description='NotFound', response='not_found')
 
         user_record.serialize(body)
         user_record.update_user()
@@ -94,9 +104,12 @@ class AdminService:
             'b': user
         }
 
-    def disabled_user(self, uid):
+    def disabled(self, uid):
         user_record = UserRecord.get_user(uid, app=admin_sdk)
         user = Admin.query.filter_by(uid=user_record.uid).first()
+
+        if not user_record or not user:
+            abort(404, description='NotFound', response='not_found')
 
         user_record.serialize({ 'disabled': not user_record.disabled })
         user_record.update_user()
@@ -111,9 +124,12 @@ class AdminService:
         }
 
 
-    def delete_user(self, uid):
+    def delete(self, uid):
         user_record = UserRecord.get_user(uid, app=admin_sdk)
         user = Admin.query.filter_by(uid=user_record.uid).first()
+
+        if not user_record or not user:
+            abort(404, description='NotFound', response='not_found')
 
         user_record.delete_user()
         user.delete()
@@ -121,7 +137,4 @@ class AdminService:
         return {
             'uid': user_record.uid
         }
-
-
-
